SPECTQL foreach notice

+++ controllers/SPECTQLController.class.php
@@ -112,10 +112,12 @@ class SPECTQLController extends AController {
 
         $tmpdir = getcwd() . "\\" .  "tmp\\*";
         $files = glob($tmpdir); // get all file names
-        foreach ($files as $file) { // iterate files
-            
-            if (is_file($file))
-                unlink($file); // delete file
+        if(isset($files)){
+	        foreach ($files as $file) { // iterate files
+	            
+	            if (is_file($file))
+	                unlink($file); // delete file
+        }
         }
     }
